add quiz for quizzing on words with less than 50% accuracy

+++ application/models.py
@@ -6,9 +6,9 @@ class Words(models.Model):
     english_word = models.CharField(max_length=40)
     spanish_word = models.CharField(max_length=40)
     category = models.CharField(max_length=40, null=True)
-    tries = models.IntegerField(default=0)
-    wrong_guesses = models.IntegerField(default=0)
-    correct_guesses = models.IntegerField(default=0)
+    tries = models.FloatField(default=0)
+    wrong_guesses = models.FloatField(default=0)
+    correct_guesses = models.FloatField(default=0)
 
 # class User(models.Model):
 #     id = models.BigAutoField(db_index=True, primary_key=True)
